generalize get_deployment -> get_configurable
