+++ pangeo_forge_orchestrator/routers/github_app.py
@@ -18,8 +18,7 @@ from gidgethub.apps import get_installation_access_token
 from sqlalchemy.orm.exc import MultipleResultsFound, NoResultFound
 from sqlmodel import Session, SQLModel, select
 
-from ..configurables.deployment import get_deployment as get_config
-from ..configurables.github_app import GitHubApp
+from ..configurables import Deployment, GitHubApp, get_configurable
 from ..dependencies import get_session as get_database_session
 from ..http import http_session
 from ..logging import logger
@@ -56,7 +55,7 @@ def html_url_to_repo_full_name(html_url: str) -> str:
 def get_jwt() -> str:
     """Adapted from https://github.com/Mariatta/gh_app_demo"""
 
-    github_app = get_config(configurable=GitHubApp)
+    github_app = get_configurable(configurable=GitHubApp)
     payload = {
         "iat": int(time.time()),
         "exp": int(time.time()) + (10 * 60),
@@ -66,7 +65,7 @@ def get_jwt() -> str:
 
 
 async def get_access_token(gh: GitHubAPI) -> str:
-    github_app = get_config(configurable=GitHubApp)
+    github_app = get_configurable(configurable=GitHubApp)
     async for installation in gh.getiter("/app/installations", jwt=get_jwt(), accept=ACCEPT):
         installation_id = installation["id"]
         # Even if installed on multiple repos within the account, I believe installations are
@@ -172,7 +171,7 @@ def get_storage_subpath_identifier(feedstock_spec: str, recipe_run: SQLModel):
     # The traitlets config doesn't offer this much configurability in `root_path` right now,
     # so just doing this here for the moment.
 
-    app_name = get_config(configurable=GitHubApp).app_name
+    app_name = get_configurable(configurable=GitHubApp).app_name
     if recipe_run.is_test:
         prefix = f"{app_name}/test/{feedstock_spec}/recipe-run-{recipe_run.id}"
     else:
@@ -346,7 +345,7 @@ async def handle_dataflow_event(
         recipe_run.completed_at = datetime.utcnow()
 
         if recipe_run.conclusion == "success":
-            bakery_config = get_config().bakeries[bakery.name]
+            bakery_config = get_configurable(configurable=Deployment).bakeries[bakery.name]
             subpath = get_storage_subpath_identifier(feedstock.spec, recipe_run)
             root_path = bakery_config["TargetStorage"]["root_path"].format(subpath=subpath)
             recipe_run.dataset_public_url = bakery_config["TargetStorage"]["public_url"].format(  # type: ignore
@@ -554,7 +553,7 @@ async def parse_payload(request, payload_bytes, event):
 async def verify_hash_signature(request: Request, payload_bytes: bytes) -> None:
     if hash_signature := request.headers.get("X-Hub-Signature-256", None):
 
-        github_app = get_config(configurable=GitHubApp)
+        github_app = get_configurable(configurable=GitHubApp)
         webhook_secret = bytes(github_app.webhook_secret, encoding="utf-8")  # type: ignore
         h = hmac.new(webhook_secret, payload_bytes, hashlib.sha256)
         if not hmac.compare_digest(hash_signature, f"sha256={h.hexdigest()}"):
@@ -662,7 +661,7 @@ async def run(
         MODELS["bakery"].table.id == recipe_run.bakery_id
     )
     bakery = db_session.exec(statement).one()
-    bakery_config = get_config().bakeries[bakery.name]
+    bakery_config = get_configurable(configurable=Deployment).bakeries[bakery.name]
 
     subpath = get_storage_subpath_identifier(feedstock_spec, recipe_run)
     # root paths are an interesting configuration edge-case because they combine some stable
